+++ CHANGELOG.md
@@ -6,4 +6,7 @@ All Notable changes to `gent/services-opening-hours` package.
 
 ### Added
 
-[Unreleased]:
+* DMOH-3: Added the Service value object.
+* DMOH-5: Added the ServiceCollection value object.
+
+[Unreleased]: https://github.com/StadGent/php_package_services-opening-hours/compare/master...develop